# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Firat Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/internshipperiot/all": {
            "get": {
                "tags": [
                    "Staj Dönemi"
                ],
                "description": "Staj Dönemlerini Getir",
                "operationId": "9f70c0d16962d4f962fcb2aa063ec95a",
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/internshipperiot/get": {
            "get": {
                "tags": [
                    "Staj Dönemi"
                ],
                "description": "Staj Dönemini id'ye göre getir",
                "operationId": "636bcc39ce3806b4f880378328585e3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Parametre açıklaması",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/internshipperiot/create": {
            "post": {
                "tags": [
                    "Staj Dönemi"
                ],
                "description": "Staj Dönemi Oluştur",
                "operationId": "ff664534b91e9fa63fc0e2b82837c0e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "periot_start_date",
                                    "expire_date",
                                    "periot_type"
                                ],
                                "properties": {
                                    "periot_start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "date_format": "m/d/Y"
                                    },
                                    "periot_expire_date": {
                                        "type": "string",
                                        "format": "date",
                                        "date_format": "m/d/Y"
                                    },
                                    "periot_type": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/internshipperiot/update": {
            "post": {
                "tags": [
                    "Staj Dönemi"
                ],
                "description": "Staj Dönemi Güncelle",
                "operationId": "895b8902eee51b5184e80902e84f675f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "periot_start_date",
                                    "expire_date",
                                    "periot_type"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "periot_start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "date_format": "m/d/Y"
                                    },
                                    "periot_expire_date": {
                                        "type": "string",
                                        "format": "date",
                                        "date_format": "m/d/Y"
                                    },
                                    "periot_type": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/internshipperiot/delete": {
            "post": {
                "tags": [
                    "Staj Dönemi"
                ],
                "description": "Staj Dönemi Sil",
                "operationId": "0b4d37bf7936a724953e5446ec8a1420",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/user/info": {
            "get": {
                "tags": [
                    "Kullanıcı Bilgileri"
                ],
                "description": "Kullanıcı Bilgilerini Getir",
                "operationId": "204188cd36d789d9f4b46ede80c706d8",
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/user/auth/check": {
            "get": {
                "tags": [
                    "Kullanıcı Bilgileri"
                ],
                "description": "Kullanıcı Giriş Yapmış mı?",
                "operationId": "130728dd7c19d1b832749e9c516b74d1",
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        }
    }
}